Implemented a function to get notifications for a specific user
Defined a function getNotificationsList that retrieves notifications for a user based on their userId. Utilized the Notice mongoose model to find notifications where the user is a member of the team and where the notification has not been read by the user. Populated the task field with the title property. If successful, returns the retrieved notification with a status code of 201, otherwise, logs the error and returns a status code of 400 with an error message.