Refactor user schema and model
Removed the `tin` and `city` fields from the user schema as they were only required for supplier users. Instead, added conditional requirements for these fields based on the `isSupplier` flag. This makes the schema more flexible and ensures that these fields are only required when necessary. Additionally, removed the import of `mongoose` from the user model file as it's already imported in the schema file. These changes streamline the user model structure and improve code organization.